requirements.txt
numpy
torch
scipy
pandas
pyyaml 